create single Credentials class
Included redacted print methods for each credential type
